Convert sortBy/whilst/until tests to mocha
