+++ test/test-async.js
@@ -16,42 +16,6 @@ if (!Function.prototype.bind) {
     };
 }
 
-exports['sortBy'] = function(test){
-    test.expect(2);
-
-    async.sortBy([{a:1},{a:15},{a:6}], function(x, callback){
-        setTimeout(function(){callback(null, x.a);}, 0);
-    }, function(err, result){
-        test.ok(err === null, err + " passed instead of 'null'");
-        test.same(result, [{a:1},{a:6},{a:15}]);
-        test.done();
-    });
-};
-
-exports['sortBy inverted'] = function(test){
-    test.expect(1);
-
-    async.sortBy([{a:1},{a:15},{a:6}], function(x, callback){
-        setTimeout(function(){callback(null, x.a*-1);}, 0);
-    }, function(err, result){
-        test.same(result, [{a:15},{a:6},{a:1}]);
-        test.done();
-    });
-};
-
-exports['sortBy error'] = function(test){
-    test.expect(1);
-    var error = new Error('asdas');
-    async.sortBy([{a:1},{a:15},{a:6}], function(x, callback){
-        async.setImmediate(function(){
-            callback(error);
-        });
-    }, function(err){
-        test.equal(err, error);
-        test.done();
-    });
-};
-
 exports['apply'] = function(test){
     test.expect(6);
     var fn = function(){
@@ -193,210 +157,6 @@ exports['concatSeries'] = function(test){
     });
 };
 
-exports['until'] = function (test) {
-    test.expect(4);
-
-    var call_order = [];
-    var count = 0;
-    async.until(
-        function () {
-            call_order.push(['test', count]);
-            return (count == 5);
-        },
-        function (cb) {
-            call_order.push(['iteratee', count]);
-            count++;
-            cb(null, count);
-        },
-        function (err, result) {
-            test.ok(err === null, err + " passed instead of 'null'");
-            test.equals(result, 5, 'last result passed through');
-            test.same(call_order, [
-                ['test', 0],
-                ['iteratee', 0], ['test', 1],
-                ['iteratee', 1], ['test', 2],
-                ['iteratee', 2], ['test', 3],
-                ['iteratee', 3], ['test', 4],
-                ['iteratee', 4], ['test', 5],
-            ]);
-            test.equals(count, 5);
-            test.done();
-        }
-    );
-};
-
-exports['doUntil'] = function (test) {
-    test.expect(4);
-
-    var call_order = [];
-    var count = 0;
-    async.doUntil(
-        function (cb) {
-            call_order.push(['iteratee', count]);
-            count++;
-            cb(null, count);
-        },
-        function () {
-            call_order.push(['test', count]);
-            return (count == 5);
-        },
-        function (err, result) {
-            test.ok(err === null, err + " passed instead of 'null'");
-            test.equals(result, 5, 'last result passed through');
-            test.same(call_order, [
-                ['iteratee', 0], ['test', 1],
-                ['iteratee', 1], ['test', 2],
-                ['iteratee', 2], ['test', 3],
-                ['iteratee', 3], ['test', 4],
-                ['iteratee', 4], ['test', 5]
-            ]);
-            test.equals(count, 5);
-            test.done();
-        }
-    );
-};
-
-exports['doUntil callback params'] = function (test) {
-    test.expect(3);
-
-    var call_order = [];
-    var count = 0;
-    async.doUntil(
-        function (cb) {
-            call_order.push(['iteratee', count]);
-            count++;
-            cb(null, count);
-        },
-        function (c) {
-            call_order.push(['test', c]);
-            return (c == 5);
-        },
-        function (err, result) {
-            if (err) throw err;
-            test.equals(result, 5, 'last result passed through');
-            test.same(call_order, [
-                ['iteratee', 0], ['test', 1],
-                ['iteratee', 1], ['test', 2],
-                ['iteratee', 2], ['test', 3],
-                ['iteratee', 3], ['test', 4],
-                ['iteratee', 4], ['test', 5]
-            ]);
-            test.equals(count, 5);
-            test.done();
-        }
-    );
-};
-
-exports['whilst'] = function (test) {
-    test.expect(4);
-
-    var call_order = [];
-
-    var count = 0;
-    async.whilst(
-        function () {
-            call_order.push(['test', count]);
-            return (count < 5);
-        },
-        function (cb) {
-            call_order.push(['iteratee', count]);
-            count++;
-            cb(null, count);
-        },
-        function (err, result) {
-            test.ok(err === null, err + " passed instead of 'null'");
-            test.equals(result, 5, 'last result passed through');
-            test.same(call_order, [
-                ['test', 0],
-                ['iteratee', 0], ['test', 1],
-                ['iteratee', 1], ['test', 2],
-                ['iteratee', 2], ['test', 3],
-                ['iteratee', 3], ['test', 4],
-                ['iteratee', 4], ['test', 5],
-            ]);
-            test.equals(count, 5);
-            test.done();
-        }
-    );
-};
-
-exports['doWhilst'] = function (test) {
-    test.expect(4);
-    var call_order = [];
-
-    var count = 0;
-    async.doWhilst(
-        function (cb) {
-            call_order.push(['iteratee', count]);
-            count++;
-            cb(null, count);
-        },
-        function () {
-            call_order.push(['test', count]);
-            return (count < 5);
-        },
-        function (err, result) {
-            test.ok(err === null, err + " passed instead of 'null'");
-            test.equals(result, 5, 'last result passed through');
-            test.same(call_order, [
-                ['iteratee', 0], ['test', 1],
-                ['iteratee', 1], ['test', 2],
-                ['iteratee', 2], ['test', 3],
-                ['iteratee', 3], ['test', 4],
-                ['iteratee', 4], ['test', 5]
-            ]);
-            test.equals(count, 5);
-            test.done();
-        }
-    );
-};
-
-exports['doWhilst callback params'] = function (test) {
-    test.expect(3);
-    var call_order = [];
-    var count = 0;
-    async.doWhilst(
-        function (cb) {
-            call_order.push(['iteratee', count]);
-            count++;
-            cb(null, count);
-        },
-        function (c) {
-            call_order.push(['test', c]);
-            return (c < 5);
-        },
-        function (err, result) {
-            if (err) throw err;
-            test.equals(result, 5, 'last result passed through');
-            test.same(call_order, [
-                ['iteratee', 0], ['test', 1],
-                ['iteratee', 1], ['test', 2],
-                ['iteratee', 2], ['test', 3],
-                ['iteratee', 3], ['test', 4],
-                ['iteratee', 4], ['test', 5]
-            ]);
-            test.equals(count, 5);
-            test.done();
-        }
-    );
-};
-
-exports['doWhilst - error'] = function (test) {
-    test.expect(1);
-    var error = new Error('asdas');
-
-    async.doWhilst(
-        function (cb) {
-            cb(error);
-        },
-        function () {},
-        function (err) {
-            test.equal(err, error);
-            test.done();
-        }
-    );
-};
-
 exports['during'] = function (test) {
     var call_order = [];
 
@@ -488,19 +248,6 @@ exports['doDuring - error iteratee'] = function (test) {
             test.done();
         }
     );
-};
-
-exports['whilst optional callback'] = function (test) {
-    var counter = 0;
-    async.whilst(
-        function () { return counter < 2; },
-        function (cb) {
-            counter++;
-            cb();
-        }
-    );
-    test.equal(counter, 2);
-    test.done();
 };
 
 exports['ensureAsync'] = {
